refactor(tui): move App construction into app/init.rs (#3314 step 1)
Pure code motion for the App god-object split: App::new (the #[cfg(test)]
shim) and App::new_with_plugin_registry (the ~830-line constructor) move
out of the central impl App block in crates/tui/src/tui/app.rs into a new
submodule crates/tui/src/tui/app/init.rs, which holds an inherent
impl App block for initialization only.

No behavior change: the constructor body is textually identical, field
defaults are untouched, and callers (including the shared test_support
constructors) keep calling App::new / App::new_with_plugin_registry via
the same paths. Small init-adjacent helpers used elsewhere
(initial_onboarding_state, onboarding_is_workspace_trust_gate,
default_composer_arrows_scroll*, push_enabled_provider_model,
discover_cached_skills) stay in app.rs; init.rs reaches them through
'use super::*'.
